requirements.txt
fastapi
uvicorn
# docker
torch
torchvision
pillow
numpy
scikit-learn
opencv-python
python-multipart